refactor: 역할 세분화
- 커넥션 풀 관리 객체
- 풀 오브젝트(커넥션) 관리 객체
- sender
